Adds portYield_FRom_ISR to I2C_ISR
This is why it was taking way longer for I2C Transactions after switching from xTaskNotify to xTaskNotifyFromISR.
xTaskNotifyFromISR does not actually start/resume the waiting task. It just notifies the scheduler to unblock it, it starts at the next tick, when the scheduler runs.
This is where portYIELD_FROM_ISR works. This whole replacement for xTaskNotify, does take about 5us longer, I noticed. But it is what FreeRTOS docs recommend, so we'll let those 5us slide.